+++ tests/datamodules/test_datamodules.py
@@ -1,10 +1,17 @@
 import uuid
 from pathlib import Path
 
+import pytest
 import torch
 from PIL import Image
 
-from pl_bolts.datamodules import CityscapesDataModule
+from pl_bolts.datamodules import (
+    BinaryMNISTDataModule,
+    CIFAR10DataModule,
+    CityscapesDataModule,
+    FashionMNISTDataModule,
+    MNISTDataModule,
+)
 from pl_bolts.datasets.cifar10_dataset import CIFAR10
 
 
@@ -63,3 +70,24 @@ def test_cityscapes_datamodule(datadir):
     img, mask = next(iter(loader))
     assert img.size() == torch.Size([batch_size, 3, 1024, 2048])
     assert mask.size() == torch.Size([batch_size, 1024, 2048])
+
+
+@pytest.mark.parametrize("val_split, train_len", [(0.2, 48_000), (5_000, 55_000)])
+def test_vision_data_module(datadir, val_split, train_len):
+    dm = _create_dm(MNISTDataModule, datadir, val_split=val_split)
+    assert len(dm.dataset_train) == train_len
+
+
+@pytest.mark.parametrize("dm_cls", [BinaryMNISTDataModule, CIFAR10DataModule, FashionMNISTDataModule, MNISTDataModule])
+def test_data_modules(datadir, dm_cls):
+    dm = _create_dm(dm_cls, datadir)
+    loader = dm.train_dataloader()
+    img, _ = next(iter(loader))
+    assert img.size() == torch.Size([2, *dm.size()])
+
+
+def _create_dm(dm_cls, datadir, val_split=0.2):
+    dm = dm_cls(data_dir=datadir, val_split=val_split, num_workers=1, batch_size=2)
+    dm.prepare_data()
+    dm.setup()
+    return dm